+++ plugins/holland.lib.lvm/tests/xfs/__init__.py
@@ -1,4 +1,3 @@
-__test__ = False
 import sys
 import os, sys
 import shutil
@@ -8,8 +7,13 @@ from nose.tools import *
 from holland.lib.lvm.raw import *
 from tests.constants import *
 
+__test__ = False
+
 def setup():
     """Setup a simple LVM device to use"""
+
+    if not __test__:
+        return
     os.environ['PATH'] = '/sbin:/usr/sbin:' + os.environ['PATH']
     size = IMG_SIZE / 512
     img_path = os.path.join(MNT_DIR, 'test.img')
@@ -36,6 +40,9 @@ def setup():
 
 def teardown():
     """Remove the previously setup LVM"""
+
+    if not __test__:
+        return
     subprocess.call("umount %s" % MNT_DIR, shell=True)
     subprocess.call("vgremove -f %s" % TEST_VG, shell=True)
     subprocess.call('losetup -d %s' % LOOP_DEV, shell=True)
